+++ src/components/List/List.tsx
@@ -97,7 +97,7 @@ export class List<T = unknown> extends React.Component<ListProps<T>, ListState<T
     }
 
     render() {
-        const {emptyPlaceholder, virtualized, className, itemsClassName} = this.props;
+        const {emptyPlaceholder, virtualized, className, itemsClassName, qa} = this.props;
 
         const {items} = this.state;
 
@@ -106,6 +106,7 @@ export class List<T = unknown> extends React.Component<ListProps<T>, ListState<T
                 {({mobile}) => (
                     <div
                         className={b({mobile}, className)}
+                        data-qa={qa}
                         tabIndex={-1}
                         onFocus={this.handleFocus}
                         onBlur={this.handleBlur}