+++ src/components/Select/types-misc.ts
@@ -1,4 +1,3 @@
 export type SelectFilterRef = {
-    getHeight: () => number | undefined;
     focus: () => void;
 };